# docs/swagger.yaml
definitions:
...
  domain.Ticket:
    properties:
      currency:
        type: string
      event_uuid:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      sold:
        type: integer
      ticket_uuid:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/ticket:
    delete:
      parameters:
      - description: Ticket UUID
        in: query
        name: ticket_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete ticket
      tags:
      - tickets
    get:
      parameters:
      - description: Ticket UUID
        in: query
        name: ticket_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ticket
      tags:
      - tickets
    post:
      consumes:
      - application/json
      parameters:
      - description: Ticket to create
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/domain.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create ticket
      tags:
      - tickets
...
swagger: "2.0"